refactor(rest-controller-test): 추상 클래스로 컨트롤러 중복 코드 제거
