+++ src/test/kotlin/apply/ui/api/CheaterRestControllerTest.kt
@@ -3,71 +3,31 @@ package apply.ui.api
 import apply.application.ApplicantService
 import apply.application.CheaterResponse
 import apply.application.CheaterService
-import apply.config.RestDocsConfiguration
 import apply.createApplicant
 import apply.domain.cheater.Cheater
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.ninjasquad.springmockk.MockkBean
 import io.mockk.every
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.FilterType
-import org.springframework.context.annotation.Import
-import org.springframework.restdocs.RestDocumentationContextProvider
-import org.springframework.restdocs.RestDocumentationExtension
-import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation
-import org.springframework.test.context.junit.jupiter.SpringExtension
-import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.delete
 import org.springframework.test.web.servlet.get
 import org.springframework.test.web.servlet.post
-import org.springframework.test.web.servlet.result.MockMvcResultHandlers
-import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder
-import org.springframework.test.web.servlet.setup.MockMvcBuilders
-import org.springframework.web.context.WebApplicationContext
-import org.springframework.web.filter.CharacterEncodingFilter
 import support.createLocalDateTime
-import support.test.TestEnvironment
 
 @WebMvcTest(
     controllers = [CheaterRestController::class],
     includeFilters = [
-        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.security.*"]),
-        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.config.*"])
+        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.security.*"])
     ]
 )
-@Import(RestDocsConfiguration::class)
-@ExtendWith(RestDocumentationExtension::class, SpringExtension::class)
-@TestEnvironment
-internal class CheaterRestControllerTest(
-    private val objectMapper: ObjectMapper,
-) {
+internal class CheaterRestControllerTest : RestControllerTest() {
     @MockkBean
-    private lateinit var applicantService: ApplicantService // todo: 이 객체를 MockBean으로 주입해야하는 이유 찾기
+    private lateinit var applicantService: ApplicantService
 
     @MockkBean
     private lateinit var cheaterService: CheaterService
-
-    private lateinit var mockMvc: MockMvc
-
-    @BeforeEach
-    internal fun setUp(
-        webApplicationContext: WebApplicationContext,
-        restDocumentationContextProvider: RestDocumentationContextProvider,
-    ) {
-        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-            .addFilter<DefaultMockMvcBuilder>(CharacterEncodingFilter("UTF-8", true))
-            .alwaysDo<DefaultMockMvcBuilder>(MockMvcResultHandlers.print())
-            .apply<DefaultMockMvcBuilder>(
-                MockMvcRestDocumentation.documentationConfiguration(
-                    restDocumentationContextProvider
-                )
-            )
-            .build()
-    }
 
     @Test
     fun `모든 부정행위자를 찾는다`() {
