+++ src/test/kotlin/apply/ui/api/EvaluationRestControllerTest.kt
@@ -4,72 +4,32 @@ import apply.application.ApplicantService
 import apply.application.EvaluationData
 import apply.application.EvaluationResponse
 import apply.application.EvaluationService
-import apply.config.RestDocsConfiguration
 import apply.createEvaluation
 import apply.createRecruitment
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.ninjasquad.springmockk.MockkBean
 import io.mockk.every
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.FilterType
-import org.springframework.context.annotation.Import
 import org.springframework.http.MediaType
-import org.springframework.restdocs.RestDocumentationContextProvider
-import org.springframework.restdocs.RestDocumentationExtension
-import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation
-import org.springframework.test.context.junit.jupiter.SpringExtension
-import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.delete
 import org.springframework.test.web.servlet.get
 import org.springframework.test.web.servlet.post
-import org.springframework.test.web.servlet.result.MockMvcResultHandlers
-import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder
-import org.springframework.test.web.servlet.setup.MockMvcBuilders
-import org.springframework.web.context.WebApplicationContext
-import org.springframework.web.filter.CharacterEncodingFilter
-import support.test.TestEnvironment
 
 @WebMvcTest(
     controllers = [EvaluationRestController::class],
     includeFilters = [
-        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.security.*"]),
-        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.config.*"])
+        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.security.*"])
     ]
 )
-@Import(RestDocsConfiguration::class)
-@ExtendWith(RestDocumentationExtension::class, SpringExtension::class)
-@TestEnvironment
-internal class EvaluationRestControllerTest(
-    private val objectMapper: ObjectMapper,
-) {
+internal class EvaluationRestControllerTest : RestControllerTest() {
 
     @MockkBean
     private lateinit var applicantService: ApplicantService
 
     @MockkBean
     private lateinit var evaluationService: EvaluationService
-
-    private lateinit var mockMvc: MockMvc
-
-    @BeforeEach
-    internal fun setUp(
-        webApplicationContext: WebApplicationContext,
-        restDocumentationContextProvider: RestDocumentationContextProvider,
-    ) {
-        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-            .addFilter<DefaultMockMvcBuilder>(CharacterEncodingFilter("UTF-8", true))
-            .alwaysDo<DefaultMockMvcBuilder>(MockMvcResultHandlers.print())
-            .apply<DefaultMockMvcBuilder>(
-                MockMvcRestDocumentation.documentationConfiguration(
-                    restDocumentationContextProvider
-                )
-            )
-            .build()
-    }
 
     @Test
     internal fun `평가를 추가한다`() {
