+++ src/test/kotlin/apply/ui/api/EvaluationTargetRestControllerTest.kt
@@ -8,58 +8,29 @@ import apply.application.EvaluationTargetData
 import apply.application.EvaluationTargetResponse
 import apply.application.EvaluationTargetService
 import apply.application.GradeEvaluationResponse
-import apply.config.RestDocsConfiguration
 import apply.createEvaluationItem
 import apply.domain.evaluationtarget.EvaluationAnswers
 import apply.domain.evaluationtarget.EvaluationStatus
 import apply.security.JwtTokenProvider
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.ninjasquad.springmockk.MockkBean
 import io.mockk.every
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
-import org.springframework.context.annotation.ComponentScan
-import org.springframework.context.annotation.FilterType
-import org.springframework.context.annotation.Import
-import org.springframework.restdocs.RestDocumentationContextProvider
-import org.springframework.restdocs.RestDocumentationExtension
-import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation
 import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders
-import org.springframework.restdocs.operation.preprocess.OperationRequestPreprocessor
-import org.springframework.restdocs.operation.preprocess.OperationResponsePreprocessor
-import org.springframework.restdocs.operation.preprocess.Preprocessors
 import org.springframework.restdocs.payload.JsonFieldType
 import org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath
 import org.springframework.restdocs.payload.PayloadDocumentation.requestFields
 import org.springframework.restdocs.payload.PayloadDocumentation.responseFields
 import org.springframework.restdocs.request.RequestDocumentation.parameterWithName
 import org.springframework.restdocs.request.RequestDocumentation.pathParameters
-import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.result.MockMvcResultHandlers
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.content
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
-import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder
-import org.springframework.test.web.servlet.setup.MockMvcBuilders
-import org.springframework.web.context.WebApplicationContext
-import org.springframework.web.filter.CharacterEncodingFilter
-import support.test.TestEnvironment
 
 @WebMvcTest(
-    controllers = [EvaluationTargetRestController::class],
-    includeFilters = [
-        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.security.*"]),
-        ComponentScan.Filter(type = FilterType.REGEX, pattern = ["apply.config.*"])
-    ]
+    controllers = [EvaluationTargetRestController::class]
 )
-@Import(RestDocsConfiguration::class)
-@ExtendWith(RestDocumentationExtension::class)
-@TestEnvironment
-internal class EvaluationTargetRestControllerTest(
-    private val objectMapper: ObjectMapper
-) {
+internal class EvaluationTargetRestControllerTest : RestControllerTest() {
     @MockkBean
     private lateinit var jwtTokenProvider: JwtTokenProvider
 
@@ -68,8 +39,6 @@ internal class EvaluationTargetRestControllerTest(
 
     @MockkBean
     private lateinit var evaluationTargetService: EvaluationTargetService
-
-    private lateinit var mockMvc: MockMvc
 
     private val recruitmentId = 1L
     private val evaluationId = 1L
@@ -98,22 +67,6 @@ internal class EvaluationTargetRestControllerTest(
         evaluationTarget = EvaluationTargetData(),
         evaluationItems = listOf(EvaluationItemResponse(createEvaluationItem()))
     )
-
-    @BeforeEach
-    internal fun setUp(
-        webApplicationContext: WebApplicationContext,
-        restDocumentationContextProvider: RestDocumentationContextProvider
-    ) {
-        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-            .addFilter<DefaultMockMvcBuilder>(CharacterEncodingFilter("UTF-8", true))
-            .alwaysDo<DefaultMockMvcBuilder>(MockMvcResultHandlers.print())
-            .apply<DefaultMockMvcBuilder>(
-                MockMvcRestDocumentation.documentationConfiguration(
-                    restDocumentationContextProvider
-                )
-            )
-            .build()
-    }
 
     @Test
     fun `평가 id와 키워드(이름, 이메일)로 평가 대상자를 조회한다`() {
@@ -274,20 +227,6 @@ internal class EvaluationTargetRestControllerTest(
                     )
                 )
             )
-    }
-
-    private fun getDocumentRequest(): OperationRequestPreprocessor {
-        return Preprocessors.preprocessRequest(
-            Preprocessors.modifyUris()
-                .scheme("https")
-                .host("woowa")
-                .removePort(),
-            Preprocessors.prettyPrint()
-        )
-    }
-
-    private fun getDocumentResponse(): OperationResponsePreprocessor {
-        return Preprocessors.preprocessResponse(Preprocessors.prettyPrint())
     }
 
     companion object {
