+++ src/test/kotlin/apply/ui/api/RestControllerTest.kt
@@ -1,0 +1,62 @@
+package apply.ui.api
+
+import apply.config.RestDocsConfiguration
+import com.fasterxml.jackson.databind.ObjectMapper
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.extension.ExtendWith
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.context.annotation.Import
+import org.springframework.restdocs.RestDocumentationContextProvider
+import org.springframework.restdocs.RestDocumentationExtension
+import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation
+import org.springframework.restdocs.operation.preprocess.OperationRequestPreprocessor
+import org.springframework.restdocs.operation.preprocess.OperationResponsePreprocessor
+import org.springframework.restdocs.operation.preprocess.Preprocessors
+import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.result.MockMvcResultHandlers
+import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder
+import org.springframework.test.web.servlet.setup.MockMvcBuilders
+import org.springframework.web.context.WebApplicationContext
+import org.springframework.web.filter.CharacterEncodingFilter
+import support.test.TestEnvironment
+
+@Import(RestDocsConfiguration::class)
+@ExtendWith(RestDocumentationExtension::class)
+@TestEnvironment
+abstract class RestControllerTest {
+
+    @Autowired
+    lateinit var objectMapper: ObjectMapper
+
+    lateinit var mockMvc: MockMvc
+
+    @BeforeEach
+    internal fun setUp(
+        webApplicationContext: WebApplicationContext,
+        restDocumentationContextProvider: RestDocumentationContextProvider
+    ) {
+        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
+            .addFilter<DefaultMockMvcBuilder>(CharacterEncodingFilter("UTF-8", true))
+            .alwaysDo<DefaultMockMvcBuilder>(MockMvcResultHandlers.print())
+            .apply<DefaultMockMvcBuilder>(
+                MockMvcRestDocumentation.documentationConfiguration(
+                    restDocumentationContextProvider
+                )
+            )
+            .build()
+    }
+
+    fun getDocumentRequest(): OperationRequestPreprocessor {
+        return Preprocessors.preprocessRequest(
+            Preprocessors.modifyUris()
+                .scheme("https")
+                .host("woowa")
+                .removePort(),
+            Preprocessors.prettyPrint()
+        )
+    }
+
+    fun getDocumentResponse(): OperationResponsePreprocessor {
+        return Preprocessors.preprocessResponse(Preprocessors.prettyPrint())
+    }
+}
